refactor: eliminate redundant prompt helper files
Merge promptHelpers.ts and toolUsePrompt.ts into PromptBuilder.ts:
- Inline getSystemPromptWithRules() function
- Inline TOOL_USE_INSTRUCTIONS constant

Reduces prompt utilities from 4 files to 2:
- promptUtils.ts: Core utilities (XML, rendering, file ops)
- PromptBuilder.ts: Prompt construction logic + helpers
